improved test for Ogg Vorbis support
This is based on a series of assumptions:
* If the audio tag has a volume attribute; and
* If the browser is Firefox

the we assume it can play Ogg Vorbis.

This gets arround the current problem of Safari having audio tag support, but only for
types supported by Quicktime. The previous test gave a false postive for Ogg
(based on using the volume attribute only).

When the canPlayType() function is landed in Firefox and WebKit then we'll
use that instead.