added code to delay the execution of commands until the flash player has loaded.
Even though the flash object was available in the DOM this did aways
means that the flash player could accept commands.

Presumably this was because of the initialisation time in the player itself.

The SWF var is not made available until the onFlashLoad callback is
made from the flash movie, which means that any calls made against
SWF before this fail.

This patch added a check that queues the commands if flash is not
ready, and runs them when it is.

This makes the player work correctly when a load and play command
are run on DOM ready.