Moved code to fix flash time reporting to flash command handling
The flash player does not reset the position of a clip to 0
when stop is used.

The code to work-around this was in the main player stop function
when really it should be in the flash commmand processing section.

This patch moves the code into the flash command processor
(It could probably be moved into the flash movie)