fixed a bug where loading no clip sent NaN in the current position at startup
This was because with no clip loaded the audioCommand did
not return a value (as the file name match failed for both types)

The current position code has been move to it own function that
return 0 when the player is stopped. Essentially the same code was
moved out of the elapsedTime command for each type. Both these
commamds now return the raw current time for the clip. The
current_position() calls now calls these if the played is
not stopped.