added a cookie to signal the state of the popup player window
Set a cookie so that the parent window knows that we are open.

We have to signal via cookies as the user might have moved to another
page in the parent window since the player was opened.

This is a valid action - the user might be listening to  audio, and we
don't want to reload new content into the same window if they click
on another link but instead pass the new URL via a function call.

Apart from saving on traffic, this will allow a future playlist
functionality to work correctly.