added code to manage the popup from the onclick event
Obviously this is trickier than a normal app, where the user will
stay on one page with the child window open.

In this case audio could be playing for the chold window while the
user moves around the site.

This patch allows the player window to stay open and for messages
to be sent to it either directly (via the normal parent/child link)
or via a cookie if the p/c link has been broken by the user going
to another page.

We do this so that the player can remain open and playing audio
and so that new items can be added without interupting playback.