added cookie playlist to popup module
The cookie playlist provides a framework for the onclick event
to store unique ids in a cookie.

When an item is clicked a unique id is grabbed from the elements id
attribute and passed into the onclick handler

This id is then pushed onto the playlist (or ignored if it is already
in the list).

The cookie will be accessed from the player window, either as part of
a timed event, or a function call if it is available