refactored cookie playlist plugin to return a cookie object
Sometimes two playlists are needed, so we need an object for each
to ensure that that the two don't get mixed up!

In this case one is a queue for passing items between windows,
and the other is the real playlist i.e. the items that are
showing in the user playlist.