replace function call to child player window with cookie queue
A direct call does not work because the player is not ready when it has
first been created. It could have been done by URL for the intitial call,
then the function call, and finally via a cookie

It seemed silly to have three methods to get data to the player, each used
depending on the state of the created window.

The cookie based queue seems adequate at this point, even without some
sort of locking mechanism. A cookie-based queue with locking is probably
over-kill.