added complete example-02
This example passes the id of the audio link to the player in the
expectation that the player will do something with it. For example
it could go and grab some info from the server (such as a file to play)
and fill our the title and description areas.

More work on this example