on position change event now includes data to position a slider
The event includes per mille data (1/1000) that can be picked up
and used to drive a slider that show the audio position.

A new duration audio command was added the returns the duration
of the clip and this is used by the update_position call to
calculate the per mille value (current position / duration) * 1000

per mille was used instead of percent to ensure the slider moves
smoothly, and the update default value was reduced to 0.1 seconds
(from 1.0 seconds).