updated example-01 to use a slider to show position in the clip
This uses the unobtrusive accessable slider from frequency-decoder
to show the current position in the clip. The slider is driven by
events from the audioPlayer plugin and works for both Vorbis and
MP3 files.

This particular slider was chosen because of its attention to
accessibility details.

Later this slider will be used for full interactive audio seek.