added new methods to fdslider to allow dynamic control of some value
This ones we can control now are:
* max
This is to allow the dynamic changing of the slider length and
number of steps without having to recreate it. The length can be
set (in our case) to the number of seconds in the item.

* slider position

the slider postion is changed to the value specified. As with above
this allows it to be driven to sync with the time readout.

Both these were added to avoid having to scale between time and
percentage or permille and back again.

The side effect of these changes is that key-based nudging now makes
sense - each key press can correspond to a set number of seconds
rather than some in between value. For accessibilty this makes more
sense.