added metadata structure to audio player module and load function
This is to allow the data extracted from a remote database to be used
to drive the player. The data structure contains all the information
required by the player, so there is no need for it to get the metadata
from the audio itself

This makes the player faster to load up as it does not have to wait for
the audio files to load.

The metadata also provides mp3 and ogg file and eventually this will
allow for a default playback scheme:

* play vorbis if we can and there is a file
* fallback to mp3 otherwise

... or perhaps some other schemes.