changed fd slider so keyboard nudging of the slider can have their own increment values
The slider increment can be changed via a class on the input field, with ctrl key
providing a 2x multiplier.

The keyboard nudging of the slider used these same values. This change allows
different values to be set for the keyboard via the fd_kbinc and
fd_kbhop classes.

This is useful as mouse control of the slider can continue to be single steps
while the keyboard can be used to get bigger hops.

In practice this is used when the increments are in seconds. Keyboard
nudge can be set to fixed values of 5 and 30 seconds to make it
simpler for keyboard users (screen readers) to move around.