fixed bug where slider jumped back to the old postion momentarily when seeking with ogg files
This was because the <audio> element IS seeking when the
location is changed and so reports the old position back
to the slider and the timer when play restarts (and before
it has seeked to the new location).

Fixed by seperating out tracking of a seek on the slider and
the actual seek of the audio element, and toggling a value depending
on whether <audio> is 'seeking' or has 'seeked'.

If this value is true (audio is still seeking) the slider and timer
do not update.