first stage of seperating out the player objects and the interface
The player objects - flash and audio tag - have been extracted into
seperate objects which are created in the mediaplayer and used as
required.

This was done to make it simpler to add new ones (like video tags)
and abstract them in such as way as the interfacing is the same
for all types.

In the future this will mean the media player can be used for any
kind of content. This is useful because the final version of the
player wil be used to play Vorbis files and fallback to MP3 if
the browser supports it.

This logic might go like this:

if ( mediaPlayer.canPlay('vorbis')){
   mediaPlayer.load( vorbis file );
}
else{
  mediaPlayer.load( mp3 file );
}

and then

mediaPlayer.play();
mediaPlayer.pause();

etc