moved tracking of current state from player objects into mediaplayer plugin
This was done so that the media objects just have the bare
minimum of tracking required to make then all behave the same
when accessed from the mediaplayer.

For example, the MP3 object tracks the current url because it has to
be passed to each function. The audio tag object does not.