added media_type attribute to load function and video tag support
It is now possible to give the load function am Ogg Theora file
and for this to be controlled by the payer controls.

It is also possible to specify and existing object to attach the
player controls to.

Added video attach example file that shows how to attach to an
existing element.