added support for HTML5 canPlayType and did some refactoring
The code to test for codec support has been moved into a function
to make it simpler to keep up with changes in the draft and also
provide a fallback method to check in case people with older nightly
builds a using the code. The fallback test will be removed when
FF 3.1 is released.

The initialisation code of the html5_player object was also cleaned
up so that it is clearer what is going on.