fixed bug where <audio> and <video> seeked after pause.
On flash we emulate pause on by storing the time paused at, and
playing from there on resume. This is not needed for the two
HTML5 media elements as they maintain their own state.

The media object for these now tracks the pause state so as
to NOT seek after the user has paused. The exception is on a
slider initiated seek. The player must pause while the seek is
taking place, but we want to resume at the new location, so
a new param was added to pause - pausing_to_seek.

This is set to true from a slider seek so the element can
resume at the new point.

Hat tip: Bug noticed by Chris Double.