complete refactoring and addition of new features to eventwatcher
The watcher is now atttached to a specified element like other
jQuery functions.

You specify the id of the element you want to stick the watcher output
inside.

As well as the existing table counting events, there is now a scrolling
box with events in the order they are fired and the beginnings of
tracking some attributes that change.
.